Module google_cloudbuild1::api

Structs

ArtifactObjects
Files in the workspace to upload to Cloud Storage upon successful completion of all build steps.

Artifacts
Artifacts produced by a build that should be uploaded upon successful completion of all build steps.

Build
A build resource in the Cloud Build API. At a high level, a Build describes where to find source code, how to build it (for example, the builder image to run on the source), and where to store the built artifacts. Fields can include the following variables, which will be expanded when the build is created: - $PROJECT_ID: the project ID of the build. - $PROJECT_NUMBER: the project number of the build. - $BUILD_ID: the autogenerated ID of the build. - $REPO_NAME: the source repository name specified by RepoSource. - $BRANCH_NAME: the branch name specified by RepoSource. - $TAG_NAME: the tag name specified by RepoSource. - $REVISION_ID or $COMMIT_SHA: the commit SHA specified by RepoSource or resolved from the specified branch or tag. - $SHORT_SHA: first 7 characters of $REVISION_ID or $COMMIT_SHA.

BuildOptions
Optional arguments to enable specific features of builds.

BuildStep
A step in the build pipeline.

BuildTrigger
Configuration for an automated build in response to source repository changes.

BuiltImage
An image built by the pipeline.

CancelBuildRequest
Request to cancel an ongoing build.

CancelOperationRequest
The request message for Operations.CancelOperation.

CloudBuild
Central instance to access all CloudBuild related resource activities

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

FileHashes
Container message for hashes of byte content of files, used in SourceProvenance messages to verify integrity of source input to the build.

GitHubEventsConfig
GitHubEventsConfig describes the configuration of a trigger that creates a build whenever a GitHub event is received. This message is experimental.

Hash
Container message for hash values.

HttpBody
Message that represents an arbitrary HTTP body. It should only be used for payload formats that can't be represented as JSON, such as raw binary or an HTML page. This message can be used both in streaming and non-streaming API methods in the request as well as the response. It can be used as a top-level request field, which is convenient if one wants to extract parameters from either the URL or HTTP template into the request fields and also want access to the raw HTTP body. Example: message GetResourceRequest { // A unique request id. string request_id = 1; // The raw HTTP body is bound to this field. google.api.HttpBody http_body = 2; } service ResourceService { rpc GetResource(GetResourceRequest) returns (google.api.HttpBody); rpc UpdateResource(google.api.HttpBody) returns (google.protobuf.Empty); } Example with streaming methods: service CaldavService { rpc GetCalendar(stream google.api.HttpBody) returns (stream google.api.HttpBody); rpc UpdateCalendar(stream google.api.HttpBody) returns (stream google.api.HttpBody); } Use of this type only changes how the request and response bodies are handled, all other features will continue to work unchanged.

InlineSecret
Pairs a set of secret environment variables mapped to encrypted values with the Cloud KMS key to use to decrypt the value.

ListBuildTriggersResponse
Response containing existing BuildTriggers.

ListBuildsResponse
Response including listed builds.

Operation
This resource represents a long-running operation that is the result of a network API call.

OperationCancelCall
Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

OperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

OperationMethods
A builder providing access to all methods supported on operation resources. It is not used directly, but through the CloudBuild hub.

ProjectBuildCancelCall
Cancels a build in progress.

ProjectBuildCreateCall
Starts a build with the specified configuration. This method returns a long-running Operation, which includes the build ID. Pass the build ID to GetBuild to determine the build status (such as SUCCESS or FAILURE).

ProjectBuildGetCall
Returns information about a previously requested build. The Build that is returned includes its status (such as SUCCESS, FAILURE, or WORKING), and timing information.

ProjectBuildListCall
Lists previously requested builds. Previously requested builds may still be in-progress, or may have finished successfully or unsuccessfully.

ProjectBuildRetryCall
Creates a new build based on the specified build. This method creates a new build using the original build request, which may or may not result in an identical build. For triggered builds: * Triggered builds resolve to a precise revision; therefore a retry of a triggered build will result in a build that uses the same revision. For non-triggered builds that specify RepoSource: * If the original build built from the tip of a branch, the retried build will build from the tip of that branch, which may not be the same revision as the original build. * If the original build specified a commit sha or revision ID, the retried build will use the identical source. For builds that specify StorageSource: * If the original build pulled source from Google Cloud Storage without specifying the generation of the object, the new build will use the current object, which may be different from the original build source. * If the original build pulled source from Cloud Storage and specified the generation of the object, the new build will attempt to use the same object, which may or may not be available depending on the bucket's lifecycle management settings.

ProjectLocationBuildCancelCall
Cancels a build in progress.

ProjectLocationBuildCreateCall
Starts a build with the specified configuration. This method returns a long-running Operation, which includes the build ID. Pass the build ID to GetBuild to determine the build status (such as SUCCESS or FAILURE).

ProjectLocationBuildGetCall
Returns information about a previously requested build. The Build that is returned includes its status (such as SUCCESS, FAILURE, or WORKING), and timing information.

ProjectLocationBuildListCall
Lists previously requested builds. Previously requested builds may still be in-progress, or may have finished successfully or unsuccessfully.

ProjectLocationBuildRetryCall
Creates a new build based on the specified build. This method creates a new build using the original build request, which may or may not result in an identical build. For triggered builds: * Triggered builds resolve to a precise revision; therefore a retry of a triggered build will result in a build that uses the same revision. For non-triggered builds that specify RepoSource: * If the original build built from the tip of a branch, the retried build will build from the tip of that branch, which may not be the same revision as the original build. * If the original build specified a commit sha or revision ID, the retried build will use the identical source. For builds that specify StorageSource: * If the original build pulled source from Google Cloud Storage without specifying the generation of the object, the new build will use the current object, which may be different from the original build source. * If the original build pulled source from Cloud Storage and specified the generation of the object, the new build will attempt to use the same object, which may or may not be available depending on the bucket's lifecycle management settings.

ProjectLocationOperationCancelCall
Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

ProjectLocationOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the CloudBuild hub.

ProjectTriggerCreateCall
Creates a new BuildTrigger. This API is experimental.

ProjectTriggerDeleteCall
Deletes a BuildTrigger by its project ID and trigger ID. This API is experimental.

ProjectTriggerGetCall
Returns information about a BuildTrigger. This API is experimental.

ProjectTriggerListCall
Lists existing BuildTriggers. This API is experimental.

ProjectTriggerPatchCall
Updates a BuildTrigger by its project ID and trigger ID. This API is experimental.

ProjectTriggerRunCall
Runs a BuildTrigger at a particular source revision.

ProjectTriggerWebhookCall
ReceiveTriggerWebhook [Experimental] is called when the API receives a webhook request targeted at a specific trigger.

PullRequestFilter
PullRequestFilter contains filter properties for matching GitHub Pull Requests.

PushFilter
Push contains filter properties for matching GitHub git pushes.

ReceiveTriggerWebhookResponse
ReceiveTriggerWebhookResponse [Experimental] is the response object for the ReceiveTriggerWebhook method.

RepoSource
Location of the source in a Google Cloud Source Repository.

Results
Artifacts created by the build pipeline.

RetryBuildRequest
Specifies a build to retry.

Secret
Pairs a set of secret environment variables containing encrypted values with the Cloud KMS key to use to decrypt the value. Note: Use kmsKeyName with available_secrets instead of using kmsKeyName with secret. For instructions see: https://cloud.google.com/cloud-build/docs/securing-builds/use-encrypted-credentials.

SecretManagerSecret
Pairs a secret environment variable with a SecretVersion in Secret Manager.

Secrets
Secrets and secret environment variables.

Source
Location of the source in a supported storage service.

SourceProvenance
Provenance of the source. Ways to find the original source, or verify that some source was used for this build.

Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.

StorageSource
Location of the source in an archive file in Google Cloud Storage.

StorageSourceManifest
Location of the source manifest in Google Cloud Storage. This feature is in Preview.

TimeSpan
Start and end times for a build execution phase.

Volume
Volume describes a Docker container volume which is mounted into build steps in order to persist files across build step execution.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.