Struct google_cloudbuild1::api::Build

pub struct Build {
    pub artifacts: Option<Artifacts>,
    pub available_secrets: Option<Secrets>,
    pub build_trigger_id: Option<String>,
    pub create_time: Option<String>,
    pub finish_time: Option<String>,
    pub id: Option<String>,
    pub images: Option<Vec<String>>,
    pub log_url: Option<String>,
    pub logs_bucket: Option<String>,
    pub name: Option<String>,
    pub options: Option<BuildOptions>,
    pub project_id: Option<String>,
    pub queue_ttl: Option<String>,
    pub results: Option<Results>,
    pub secrets: Option<Vec<Secret>>,
    pub service_account: Option<String>,
    pub source: Option<Source>,
    pub source_provenance: Option<SourceProvenance>,
    pub start_time: Option<String>,
    pub status: Option<String>,
    pub status_detail: Option<String>,
    pub steps: Option<Vec<BuildStep>>,
    pub substitutions: Option<HashMap<String, String>>,
    pub tags: Option<Vec<String>>,
    pub timeout: Option<String>,
    pub timing: Option<HashMap<String, TimeSpan>>,
}

A build resource in the Cloud Build API. At a high level, a Build describes where to find source code, how to build it (for example, the builder image to run on the source), and where to store the built artifacts. Fields can include the following variables, which will be expanded when the build is created: - $PROJECT_ID: the project ID of the build. - $PROJECT_NUMBER: the project number of the build. - $BUILD_ID: the autogenerated ID of the build. - $REPO_NAME: the source repository name specified by RepoSource. - $BRANCH_NAME: the branch name specified by RepoSource. - $TAG_NAME: the tag name specified by RepoSource. - $REVISION_ID or $COMMIT_SHA: the commit SHA specified by RepoSource or resolved from the specified branch or tag. - $SHORT_SHA: first 7 characters of $REVISION_ID or $COMMIT_SHA.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

builds cancel projects (response)

builds create projects (request)

builds get projects (response)

locations builds cancel projects (response)

locations builds create projects (request)

locations builds get projects (response)

Fields
artifacts: Option<Artifacts>

Artifacts produced by the build that should be uploaded upon successful completion of all build steps.

available_secrets: Option<Secrets>

Secrets and secret environment variables.

build_trigger_id: Option<String>

Output only. The ID of the BuildTrigger that triggered this build, if it was triggered automatically.

create_time: Option<String>

Output only. Time at which the request to create the build was received.

finish_time: Option<String>

Output only. Time at which execution of the build was finished. The difference between finish_time and start_time is the duration of the build's execution.

id: Option<String>

Output only. Unique identifier of the build.

images: Option<Vec<String>>

A list of images to be pushed upon the successful completion of all build steps. The images are pushed using the builder service account's credentials. The digests of the pushed images will be stored in the Build resource's results field. If any of the images fail to be pushed, the build status is marked FAILURE.

log_url: Option<String>

Output only. URL to logs for this build in Google Cloud Console.

logs_bucket: Option<String>

Google Cloud Storage bucket where logs should be written (see Bucket Name Requirements). Logs file names will be of the format ${logs_bucket}/log-${build_id}.txt.

name: Option<String>

Output only. The 'Build' name with format: projects/{project}/locations/{location}/builds/{build}, where {build} is a unique identifier generated by the service.

options: Option<BuildOptions>

Special options for this build.

project_id: Option<String>

Output only. ID of the project.

queue_ttl: Option<String>

TTL in queue for this build. If provided and the build is enqueued longer than this value, the build will expire and the build status will be EXPIRED. The TTL starts ticking from create_time.

results: Option<Results>

Output only. Results of the build.

secrets: Option<Vec<Secret>>

Secrets to decrypt using Cloud Key Management Service. Note: Secret Manager is the recommended technique for managing sensitive data with Cloud Build. Use available_secrets to configure builds to access secrets from Secret Manager. For instructions, see: https://cloud.google.com/cloud-build/docs/securing-builds/use-secrets

service_account: Option<String>

IAM service account whose credentials will be used at build runtime. Must be of the format projects/{PROJECT_ID}/serviceAccounts/{ACCOUNT}. ACCOUNT can be email address or uniqueId of the service account. This field is in beta.

source: Option<Source>

The location of the source files to build.

source_provenance: Option<SourceProvenance>

Output only. A permanent fixed identifier for source.

start_time: Option<String>

Output only. Time at which execution of the build was started.

status: Option<String>

Output only. Status of the build.

status_detail: Option<String>

Output only. Customer-readable message about the current status.

steps: Option<Vec<BuildStep>>

Required. The operations to be performed on the workspace.

substitutions: Option<HashMap<String, String>>

Substitutions data for Build resource.

tags: Option<Vec<String>>

Tags for annotation of a Build. These are not docker tags.

timeout: Option<String>

Amount of time that this build should be allowed to run, to second granularity. If this amount of time elapses, work on the build will cease and the build status will be TIMEOUT. timeout starts ticking from startTime. Default time is ten minutes.

timing: Option<HashMap<String, TimeSpan>>

Output only. Stores timing information for phases of the build. Valid keys are: * BUILD: time to execute all build steps * PUSH: time to push all specified images. * FETCHSOURCE: time to fetch source. If the build does not specify source or images, these keys will not be included.