Struct google_cloudbuild1::api::BuildOptions

pub struct BuildOptions {
    pub disk_size_gb: Option<String>,
    pub dynamic_substitutions: Option<bool>,
    pub env: Option<Vec<String>>,
    pub log_streaming_option: Option<String>,
    pub logging: Option<String>,
    pub machine_type: Option<String>,
    pub requested_verify_option: Option<String>,
    pub secret_env: Option<Vec<String>>,
    pub source_provenance_hash: Option<Vec<String>>,
    pub substitution_option: Option<String>,
    pub volumes: Option<Vec<Volume>>,
    pub worker_pool: Option<String>,
}

Optional arguments to enable specific features of builds.

This type is not used in any activity, and only used as part of another schema.

Fields
disk_size_gb: Option<String>

Requested disk size for the VM that runs the build. Note that this is NOT "disk free"; some of the space will be used by the operating system and build utilities. Also note that this is the minimum disk size that will be allocated for the build -- the build may run with a larger disk than requested. At present, the maximum disk size is 1000GB; builds that request more than the maximum are rejected with an error.

dynamic_substitutions: Option<bool>

Option to specify whether or not to apply bash style string operations to the substitutions. NOTE: this is always enabled for triggered builds and cannot be overridden in the build configuration file.

env: Option<Vec<String>>

A list of global environment variable definitions that will exist for all build steps in this build. If a variable is defined in both globally and in a build step, the variable will use the build step value. The elements are of the form "KEY=VALUE" for the environment variable "KEY" being given the value "VALUE".

log_streaming_option: Option<String>

Option to define build log streaming behavior to Google Cloud Storage.

logging: Option<String>

Option to specify the logging mode, which determines if and where build logs are stored.

machine_type: Option<String>

Compute Engine machine type on which to run the build.

requested_verify_option: Option<String>

Requested verifiability options.

secret_env: Option<Vec<String>>

A list of global environment variables, which are encrypted using a Cloud Key Management Service crypto key. These values must be specified in the build's Secret. These variables will be available to all build steps in this build.

source_provenance_hash: Option<Vec<String>>

Requested hash for SourceProvenance.

substitution_option: Option<String>

Option to specify behavior when there is an error in the substitution checks. NOTE: this is always set to ALLOW_LOOSE for triggered builds and cannot be overridden in the build configuration file.

volumes: Option<Vec<Volume>>

Global list of volumes to mount for ALL build steps Each volume is created as an empty volume prior to starting the build process. Upon completion of the build, volumes and their contents are discarded. Global volume names and paths cannot conflict with the volumes defined a build step. Using a global volume in a build with only one step is not valid as it is indicative of a build request with an incorrect configuration.

worker_pool: Option<String>

Option to specify a WorkerPool for the build. Format: projects/{project}/locations/{location}/workerPools/{workerPool} This field is in beta and is available only to restricted users.