Struct google_cloudbuild1::api::ProjectMethods

pub struct ProjectMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on project resources. It is not used directly, but through the CloudBuild hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_cloudbuild1 as cloudbuild1;
 
use std::default::Default;
use oauth2;
use cloudbuild1::CloudBuild;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = CloudBuild::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `builds_cancel(...)`, `builds_create(...)`, `builds_get(...)`, `builds_list(...)`, `builds_retry(...)`, `locations_builds_cancel(...)`, `locations_builds_create(...)`, `locations_builds_get(...)`, `locations_builds_list(...)`, `locations_builds_retry(...)`, `locations_operations_cancel(...)`, `locations_operations_get(...)`, `triggers_create(...)`, `triggers_delete(...)`, `triggers_get(...)`, `triggers_list(...)`, `triggers_patch(...)`, `triggers_run(...)` and `triggers_webhook(...)`
// to build up your call.
let rb = hub.projects();

Implementations

impl<'a, C> ProjectMethods<'a, C>

pub fn builds_cancel(
&self,
request: CancelBuildRequest,
project_id: &str,
id: &str
) -> ProjectBuildCancelCall<'a, C>

Create a builder to help you perform the following task:

Cancels a build in progress.

Arguments

request - No description provided.

projectId - Required. ID of the project.

id - Required. ID of the build.

pub fn builds_create(
&self,
request: Build,
project_id: &str
) -> ProjectBuildCreateCall<'a, C>

Create a builder to help you perform the following task:

Starts a build with the specified configuration. This method returns a long-running Operation, which includes the build ID. Pass the build ID to GetBuild to determine the build status (such as SUCCESS or FAILURE).

Arguments

request - No description provided.

projectId - Required. ID of the project.

pub fn builds_get(
&self,
project_id: &str,
id: &str
) -> ProjectBuildGetCall<'a, C>

Create a builder to help you perform the following task:

Returns information about a previously requested build. The Build that is returned includes its status (such as SUCCESS, FAILURE, or WORKING), and timing information.

Arguments

projectId - Required. ID of the project.

id - Required. ID of the build.

pub fn builds_list(&self, project_id: &str) -> ProjectBuildListCall<'a, C>

Create a builder to help you perform the following task:

Lists previously requested builds. Previously requested builds may still be in-progress, or may have finished successfully or unsuccessfully.

Arguments

projectId - Required. ID of the project.

pub fn builds_retry(
&self,
request: RetryBuildRequest,
project_id: &str,
id: &str
) -> ProjectBuildRetryCall<'a, C>

Create a builder to help you perform the following task:

Creates a new build based on the specified build. This method creates a new build using the original build request, which may or may not result in an identical build. For triggered builds: * Triggered builds resolve to a precise revision; therefore a retry of a triggered build will result in a build that uses the same revision. For non-triggered builds that specify RepoSource: * If the original build built from the tip of a branch, the retried build will build from the tip of that branch, which may not be the same revision as the original build. * If the original build specified a commit sha or revision ID, the retried build will use the identical source. For builds that specify StorageSource: * If the original build pulled source from Google Cloud Storage without specifying the generation of the object, the new build will use the current object, which may be different from the original build source. * If the original build pulled source from Cloud Storage and specified the generation of the object, the new build will attempt to use the same object, which may or may not be available depending on the bucket's lifecycle management settings.

Arguments

request - No description provided.

projectId - Required. ID of the project.

id - Required. Build ID of the original build.

pub fn locations_builds_cancel(
&self,
request: CancelBuildRequest,
name: &str
) -> ProjectLocationBuildCancelCall<'a, C>

Create a builder to help you perform the following task:

Cancels a build in progress.

Arguments

request - No description provided.

name - The name of the Build to cancel. Format: projects/{project}/locations/{location}/builds/{build}

pub fn locations_builds_create(
&self,
request: Build,
parent: &str
) -> ProjectLocationBuildCreateCall<'a, C>

Create a builder to help you perform the following task:

Starts a build with the specified configuration. This method returns a long-running Operation, which includes the build ID. Pass the build ID to GetBuild to determine the build status (such as SUCCESS or FAILURE).

Arguments

request - No description provided.

parent - The parent resource where this build will be created. Format: projects/{project}/locations/{location}

pub fn locations_builds_get(
&self,
name: &str
) -> ProjectLocationBuildGetCall<'a, C>

Create a builder to help you perform the following task:

Returns information about a previously requested build. The Build that is returned includes its status (such as SUCCESS, FAILURE, or WORKING), and timing information.

Arguments

name - The name of the Build to retrieve. Format: projects/{project}/locations/{location}/builds/{build}

pub fn locations_builds_list(
&self,
parent: &str
) -> ProjectLocationBuildListCall<'a, C>

Create a builder to help you perform the following task:

Lists previously requested builds. Previously requested builds may still be in-progress, or may have finished successfully or unsuccessfully.

Arguments

parent - The parent of the collection of Builds. Format: projects/{project}/locations/location

pub fn locations_builds_retry(
&self,
request: RetryBuildRequest,
name: &str
) -> ProjectLocationBuildRetryCall<'a, C>

Create a builder to help you perform the following task:

Creates a new build based on the specified build. This method creates a new build using the original build request, which may or may not result in an identical build. For triggered builds: * Triggered builds resolve to a precise revision; therefore a retry of a triggered build will result in a build that uses the same revision. For non-triggered builds that specify RepoSource: * If the original build built from the tip of a branch, the retried build will build from the tip of that branch, which may not be the same revision as the original build. * If the original build specified a commit sha or revision ID, the retried build will use the identical source. For builds that specify StorageSource: * If the original build pulled source from Google Cloud Storage without specifying the generation of the object, the new build will use the current object, which may be different from the original build source. * If the original build pulled source from Cloud Storage and specified the generation of the object, the new build will attempt to use the same object, which may or may not be available depending on the bucket's lifecycle management settings.

Arguments

request - No description provided.

name - The name of the Build to retry. Format: projects/{project}/locations/{location}/builds/{build}

pub fn locations_operations_cancel(
&self,
request: CancelOperationRequest,
name: &str
) -> ProjectLocationOperationCancelCall<'a, C>

Create a builder to help you perform the following task:

Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

Arguments

request - No description provided.

name - The name of the operation resource to be cancelled.

pub fn locations_operations_get(
&self,
name: &str
) -> ProjectLocationOperationGetCall<'a, C>

Create a builder to help you perform the following task:

Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

Arguments

name - The name of the operation resource.

pub fn triggers_create(
&self,
request: BuildTrigger,
project_id: &str
) -> ProjectTriggerCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new BuildTrigger. This API is experimental.

Arguments

request - No description provided.

projectId - Required. ID of the project for which to configure automatic builds.

pub fn triggers_delete(
&self,
project_id: &str,
trigger_id: &str
) -> ProjectTriggerDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a BuildTrigger by its project ID and trigger ID. This API is experimental.

Arguments

projectId - Required. ID of the project that owns the trigger.

triggerId - Required. ID of the BuildTrigger to delete.

pub fn triggers_get(
&self,
project_id: &str,
trigger_id: &str
) -> ProjectTriggerGetCall<'a, C>

Create a builder to help you perform the following task:

Returns information about a BuildTrigger. This API is experimental.

Arguments

projectId - Required. ID of the project that owns the trigger.

triggerId - Required. Identifier (id or name) of the BuildTrigger to get.

pub fn triggers_list(&self, project_id: &str) -> ProjectTriggerListCall<'a, C>

Create a builder to help you perform the following task:

Lists existing BuildTriggers. This API is experimental.

Arguments

projectId - Required. ID of the project for which to list BuildTriggers.

pub fn triggers_patch(
&self,
request: BuildTrigger,
project_id: &str,
trigger_id: &str
) -> ProjectTriggerPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a BuildTrigger by its project ID and trigger ID. This API is experimental.

Arguments

request - No description provided.

projectId - Required. ID of the project that owns the trigger.

triggerId - Required. ID of the BuildTrigger to update.

pub fn triggers_run(
&self,
request: RepoSource,
project_id: &str,
trigger_id: &str
) -> ProjectTriggerRunCall<'a, C>

Create a builder to help you perform the following task:

Runs a BuildTrigger at a particular source revision.

Arguments

request - No description provided.

projectId - Required. ID of the project.

triggerId - Required. ID of the trigger.

pub fn triggers_webhook(
&self,
request: HttpBody,
project_id: &str,
trigger: &str
) -> ProjectTriggerWebhookCall<'a, C>

Create a builder to help you perform the following task:

ReceiveTriggerWebhook [Experimental] is called when the API receives a webhook request targeted at a specific trigger.

Arguments

request - No description provided.

projectId - Project in which the specified trigger lives

trigger - Name of the trigger to run the payload against